+++ test/unit/specs/store/index.spec.js
@@ -1,9 +1,8 @@
 const Modules = require(`src/vuex/modules`).default
-const node = require(`../../helpers/node_mock`)
 
 describe(`Module Index`, () => {
   it(`can be instantiated`, () => {
-    const modules = Modules({ node })
+    const modules = Modules({ node: {} })
 
     // check modules for correct export interface
     for (const moduleName in modules) {
