+++ test/unit/specs/store/wallet.spec.js
@@ -1,7 +1,7 @@
 import walletModule from "modules/wallet.js"
-import lcdClientMock from "src/connectors/lcdClientMock.js"
+import mockValues from "test/unit/helpers/mockValues.js"
 
-const { stakingParameters } = lcdClientMock.state
+const { stakingParameters } = mockValues.state
 
 const mockRootState = {
   stakingParameters,
